Update install script to be more self contained
The idea is to be able to call the install script directly from the
internet. `curl -L <url>/install` or something. When sourcing another
files the process fails as it cannot find the file. By Including it in
the install file this inscreased the chance of this install script
working.