Combine zsh + bash into generic shells grouping
Mostly zsh and bash can share the same config. For better or for worse
they are in the same config folder now. The process has been good to
thin out some things which are not being used any more.

Still to do:
- Exports
- Options
- Functions

(although these might not need to be done. it's not like I actually use
 bash ever)